add EventEmitter for paginator #1

+++ core/src/main/java/com/kspt/app/models/table/MetaDataModel.java
@@ -12,6 +12,6 @@ import java.util.List;
 @NoArgsConstructor
 public class MetaDataModel {
     public String[] columns;
-    public int totalCount; //for pagination
+    public Long totalCount; //for pagination
 //    TODO add more parameters. think about it
 }
